Fix bug writing off-library mat temps
Writer would miss if the temperature was between two
temperatures, leading to a material at temperature
~800 being used with library at 900 K. Serpent doesn't
allow this, so the routine has been updated.

Now, if a material is below the lowest temperature,
a DataWarning is raised to be consistent with other
parts of the package.

Addition unit tests have been added to ensure this works