Bump SFV solver with new public sfv package
Backend has existed on the cluster for a while, but local
regression files haven't been updated. This solver
relies on scipy.sparse.linalg.eigs to avoid making a full
matrix. The regression files have been updated and show a larger
error, but this is for a trivial 8-pin problem, where it is
likely that the fission matrix is not converged. Similarly,
the ARPACK solver cannot return N modes from the fission matrix
and thus users must specify modes between [1, N-2]. This means
that the 8-node problem can only use 6 of the modes and thus
the prediction is worse.

However, this backend performs very well for large sparse problems
and has been tested on a fuel assembly bundle.