Denote CELI, RK4, and sfv.MixedRK4 integrators as experimental
Implemented and briefly tested over a weekend a while back, but
have not been thoroughly vetted. The implementation may completely
change with more research.

Each of the integrators now allow users to pass an optional
boolean ``brave``. This will suppress a new
hydep.ExperimentalIntegratorWarning from being raised, and the default
is to raise the warning. Not passing brave will still allow the
integrator to run.

The test_integrator tests have been modified to
1. Automatically suppress the warnings using recwarn fixture
2. Perform a new test that ensures the functionality of the brave
   argument